[FIX] web: search_filters: need to convert datetime to utc
When a datetime was selected in a search filter, the time wasn't
converted to UTC to the server.
So, there were unexpecting records due to the wrong query domain.

When we retrieve the value from the datepicker, we need to convert
it to UTC before sending it to the server.