[FIX] barcodes: manage several barcode widgets
related to rev 8ae79fbceccaf00dcf4ea1333399a88b0346d4d1

It was broken following the refactoring done in saas16, resulting in
barcodes being handled on all the displayed forms and it was problematic
in the case of a form dialog openned over a "fullscreen" form view.

This was particularily problematic for the scrap dialog in the
stock_barcode module.