[FIX] web: search_filters: avoid to apply UTC offset for date fields
Due to a previous commit (@cc0a959e), if a date was selected in
search filter, the time was converted to UTC to the server.
For negative offset, the search was then executed for the day before.

Now, we avoid to convert the searched date to UTC.
We only need to convert datetime fields.